+++ game/components/base/actor/s_state.cpp
@@ -1,0 +1,16 @@
+//project verse, 2017-2021
+//by jose pazos perez
+//all rights reserved uwu
+
+#include "s_state.h"
+
+#include "gui.h"
+#include "gui_types.h"
+
+using namespace Verse;
+
+void System::State::gui(Config &c, EntityID eid) {
+#ifndef DISABLE_GUI
+    
+#endif
+}