+++ game/components/base/graphics/s_animation.cpp
@@ -5,6 +5,9 @@
 #include "s_animation.h"
 #include "log.h"
 
+#include "gui.h"
+#include "gui_types.h"
+
 using namespace Verse;
 
 void System::Animation::load(EntityID eid, YAML::Node &entity, Scene *s, Config &c) {
@@ -49,4 +52,33 @@ void System::Animation::load(EntityID eid, YAML::Node &entity, Scene *s, Config 
     }
     animation->curr_key = (entity["animation"]["curr_key"]) ? entity["animation"]["curr_key"].as<str>() : animation->frames.begin()->first;
     animation->size = (entity["animation"]["size"]) ? entity["animation"]["size"].as<int>() : 1;
+}
+
+void System::Animation::save(Component::Animation *anim, str path, std::vector<str> &key) {
+    key[2] = "animation";
+    
+    key[3] = "frames";
+    key.push_back(""); key.push_back("");
+    for (std::pair<str, Component::AnimationFrame> f : anim->frames) {
+        key[4] = f.first;
+        
+        key[5] = "index";
+        Serialization::appendYAML(path, key, f.second.index, true);
+        
+        key[5] = "fps";
+        Serialization::appendYAML(path, key, (int)round(1000.0f / (float)f.second.ms[0]), true);
+    }
+    key.pop_back(); key.pop_back();
+    
+    key[3] = "curr_key";
+    Serialization::appendYAML(path, key, (str)anim->frames.begin()->first, true);
+    
+    key[3] = "size";
+    Serialization::appendYAML(path, key, (int)anim->size, true);
+}
+
+void System::Animation::gui(Config &c, EntityID eid) {
+#ifndef DISABLE_GUI
+    
+#endif
 }