+++ game/components/base/graphics/s_texture.h
@@ -14,4 +14,6 @@ namespace Verse::System::Texture
     str getCurrKey();
     
     void load(EntityID eid, YAML::Node &entity, Scene *s, Config &c);
+    void save(Component::Texture* tex, str path, std::vector<str> &key);
+    void gui(Config &c, EntityID eid);
 }