+++ game/components/base/light/s_light.h
@@ -11,4 +11,5 @@ namespace Verse::System::Light
 {
     void render(Config &c, ui8 pid);
     void load(EntityID eid, YAML::Node &entity, Scene *s, Config &c);
+    void gui(Config &c, EntityID eid);
 }