+++ game/components/extra/s_scene_transition.h
@@ -13,4 +13,5 @@ namespace Verse::System::SceneTransition
     void handle(Config &c, Scene* new_scene, Vec2 new_pos);
 
     void load(EntityID eid, YAML::Node &entity, Scene *s, Config &c);
+    void gui(Config &c, EntityID eid);
 }