+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Moved from serialization to each component
 - From passing state to actor, created state component
 - Migrated to new Rect structure
+- Now gui and serialization are handled in each component
 
 ---